requirements.txt
django==3.2.6
plotly==4.7.1
djangorestframework==3.11.0
pandas==1.0.2
numpy==1.18.2
pyarrow==0.15.1
django-cors-headers==3.2.1
gunicorn==20.0.4
python-keycloak
python-jose
django-pandas==0.6.1
psycopg2-binary==2.8.5
psycopg2==2.8.6
django-environ==0.4.5
django-postgresql
django-postgres-extra==2.0a13
django-jsonfield==1.4.1
zeep==4.0.0
django-filter==2.4.0
uritemplate==3.0.1
drf-spectacular==0.18.0
aiohttp==3.8.0
sqlalchemy==1.4.26